python-analysis/requirements.txt
pyturso
polars
streamlit
plotly
pandas
watchdog
